Move CustomGrpcServiceError
